+++ marsworks/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.3.0"
+__version__ = "0.4.0"
 from marsworks.client import *
 from marsworks.manifest import *
 from marsworks.origin import *