+++ marsworks/client.py
@@ -1,20 +1,27 @@
+from __future__ import annotations
+
 import datetime
 import io
 import os
-from typing import Union, Optional
+import warnings
+from typing import Any, Optional, Union
 
 import httpx
 
-from marsworks.origin import Rest, Camera, Rover, BadArgumentError
-from marsworks.manifest import Manifest
-from marsworks.photo import Photo
+from .origin import BadArgumentError, Camera, Rest, Rover, Serializer
+from .manifest import Manifest
+from .photo import Photo
 
 __all__ = ("Client",)
+warnings.simplefilter("always", PendingDeprecationWarning)
 
 
 class Client:
 
-    __slots__ = ("__http",)
+    __slots__ = (
+        "__http",
+        "__sprswrngs",
+    )
 
     def __init__(
         self,
@@ -30,18 +37,23 @@ class Client:
             api_key: NASA [API key](https://api.nasa.gov/). (optional)
             session: An [AsyncClient](https://www.python-httpx.org/api/#asyncclient) object. (optional)
             suppress_warnings: Whether to suppress warnings.
+
+        Warning:
+            When api_key is not passed or it is `DEMO_KEY` a warning is sent. To suppress it
+            `suppress_warnings` must be set to `True` explicitly.
         """  # noqa: E501
         self.__http: Rest = Rest(
             api_key=api_key, session=session, suppress_warnings=suppress_warnings
         )
+        self.__sprswrngs = suppress_warnings
 
-    async def __aenter__(self):
+    async def __aenter__(self) -> Client:
         return self
 
-    async def __aexit__(self, exc_type, exc_value, exc_tb):
+    async def __aexit__(self, exc_type, exc_value, exc_tb) -> None:
         await self.close()
 
-    async def get_mission_manifest(self, name: Union[str, Rover]) -> Manifest:
+    async def get_mission_manifest(self, name: Union[str, Rover]) -> Optional[Manifest]:
         """
         Gets the mission manifest of this rover.
 
@@ -55,11 +67,9 @@ class Client:
             A [Manifest](./manifest.md) object containing mission's info.
         """  # noqa: E501
         name = Rover(name.upper() if isinstance(name, str) else name)
-
-        slz = await self.__http.start(name.value)
-        mfst = await slz.manifest_content()
-
-        return mfst
+        serializer = await self.__http.start(name.value)
+        if serializer:
+            return serializer.manifest_content()
 
     async def get_photo_by_sol(
         self,
@@ -68,7 +78,7 @@ class Client:
         *,
         camera: Optional[str] = None,
         page: Optional[int] = None,
-    ) -> list:
+    ) -> Optional[list]:
         """
         Gets the photos taken by this rover on this sol.
 
@@ -87,18 +97,24 @@ class Client:
             A list of [Photo](./photo.md) objects with url and info.
         """  # noqa: E501
         name = Rover(name.upper() if isinstance(name, str) else name)
+        if camera is not None:
+            try:
+                camera = Camera(
+                    camera.upper() if isinstance(camera, str) else camera
+                ).value
+            except ValueError:
+                if not self.__sprswrngs:
+                    warnings.warn(
+                        "Invalid value was passed for camera. "
+                        "Making request without camera."
+                    )
+                camera = None
 
-        try:
-            camera = Camera(camera.upper() if isinstance(camera, str) else camera).value
-        except ValueError:
-            camera = None
-
-        slz = await self.__http.start(
+        serializer = await self.__http.start(
             name.value + "/photos", sol=sol, camera=camera, page=page
         )
-        pht = await slz.photo_content()
-
-        return pht
+        if serializer:
+            return serializer.photo_content()
 
     async def get_photo_by_earthdate(
         self,
@@ -107,7 +123,7 @@ class Client:
         *,
         camera: Optional[str] = None,
         page: Optional[int] = None,
-    ) -> list:
+    ) -> Optional[list]:
         """
         Gets the photos taken by this rover on this date.
 
@@ -126,18 +142,24 @@ class Client:
             A list of [Photo](./photo.md) objects with url and info.
         """  # noqa: E501
         name = Rover(name.upper() if isinstance(name, str) else name)
+        if camera is not None:
+            try:
+                camera = Camera(
+                    camera.upper() if isinstance(camera, str) else camera
+                ).value
+            except ValueError:
+                if not self.__sprswrngs:
+                    warnings.warn(
+                        "Invalid value was passed for camera. "
+                        "Making request without camera."
+                    )
+                camera = None
 
-        try:
-            camera = Camera(camera.upper() if isinstance(camera, str) else camera).value
-        except ValueError:
-            camera = None
-
-        slz = await self.__http.start(
+        serializer = await self.__http.start(
             name.name + "/photos", earth_date=str(earth_date), camera=camera, page=page
         )
-        pht = await slz.photo_content()
-
-        return pht
+        if serializer:
+            return serializer.photo_content()
 
     async def get_latest_photo(
         self,
@@ -145,7 +167,7 @@ class Client:
         *,
         camera: Optional[str] = None,
         page: Optional[int] = None,
-    ) -> list:
+    ) -> Optional[list]:
         """
         Gets the latest photos taken by this rover.
 
@@ -165,20 +187,26 @@ class Client:
         *Introduced in [v0.3.0](../changelog.md#v030).*
         """  # noqa: E501
         name = Rover(name.upper() if isinstance(name, str) else name)
+        if camera is not None:
+            try:
+                camera = Camera(
+                    camera.upper() if isinstance(camera, str) else camera
+                ).value
+            except ValueError:
+                if not self.__sprswrngs:
+                    warnings.warn(
+                        "Invalid value was passed for camera. "
+                        "Making request without camera."
+                    )
+                camera = None
 
-        try:
-            camera = Camera(camera.upper() if isinstance(camera, str) else camera).value
-        except ValueError:
-            camera = None
-
-        slz = await self.__http.start(
+        serializer = await self.__http.start(
             name.name + "/latest_photos", camera=camera, page=page
         )
-        pht = await slz.photo_content()
+        if serializer:
+            return serializer.photo_content()
 
-        return pht
-
-    async def read(self, photo: Photo) -> io.BytesIO:
+    async def read(self, photo: Photo) -> Optional[io.BytesIO]:
         """
         Reads the bytes of image url in photo.
 
@@ -188,17 +216,20 @@ class Client:
         Returns:
             A [BytesIO](https://docs.python.org/3/library/io.html?highlight=bytesio#io.BytesIO) object.
         """  # noqa: E501
+        if not self.__sprswrngs:
+            warnings.warn(
+                "await Client.read() will be deprecated in version 0.5.0",
+                PendingDeprecationWarning,
+            )
         if isinstance(photo, Photo):
-            bio = await self.__http.read(photo.img_src)
-
-            return bio
-
+            if photo.img_src:
+                return await self.__http.read(photo.img_src)
         else:
             raise BadArgumentError("Photo", type(photo).__name__)
 
     async def save(
         self, photo: Photo, fp: Union[str, bytes, os.PathLike, io.BufferedIOBase]
-    ) -> int:
+    ) -> Optional[int]:
         """
         Saves the image of [Photo](./photo.md) object.
 
@@ -208,22 +239,43 @@ class Client:
 
         Returns:
             Number of bytes written.
+
+        *PendingDeprecated in [0.4.0](../changelog.md#v040). Deprecated in 0.5.0.
+        Removed in 1.0.0.*
         """  # noqa: E501
+        if not self.__sprswrngs:
+            warnings.warn(
+                "await Client.save() will be deprecated in version 0.5.0",
+                PendingDeprecationWarning,
+            )
         if isinstance(photo, Photo):
-            bio = await self.__http.read(photo.img_src)
-
-            if isinstance(fp, io.IOBase) and fp.writable():
-                bw = fp.write(bio.read1())
-
-                return bw
-
-            else:
-                with open(fp, "wb") as f:
-
-                    return f.write(bio.read1())
-
+            if photo.img_src:
+                bytes_ = await self.__http.read(photo.img_src)
+                if bytes_:
+                    if isinstance(fp, io.IOBase) and fp.writable():
+                        return fp.write(bytes_.read1())
+                    else:
+                        with open(fp, "wb") as f:  # type: ignore
+                            return f.write(bytes_.read1())
         else:
             raise BadArgumentError("Photo", type(photo).__name__)
+
+    async def get_raw_response(self, path: str, **queries: Any) -> Optional[Serializer]:
+        """
+        Gets a [Serializer](./serializer.md) containing [Response](https://www.python-httpx.org/api/#response)
+        of request made to
+        API using `path` and `queries`.
+
+        Args:
+            path: The url path.
+            queries: The endpoint to which call is to be made.
+
+        Returns:
+            A [Serializer](./serializer.md) object.
+
+        *Introduced in [v0.4.0](../changelog.md#v040).*
+        """  # noqa: E501
+        return await self.__http.start(path, **queries)
 
     async def close(self) -> None:
         """