+++ marsworks/origin/serializer.py
@@ -1,48 +1,65 @@
 import inspect
+from typing import Optional
 
 import httpx
 import marsworks
-from marsworks.manifest import Manifest
-from marsworks.origin.exceptions import BadContentError
+
+from ..manifest import Manifest
+from .exceptions import BadContentError
 
 __all__ = ("Serializer",)
 
 
 class Serializer:
+    """
+    A class representing a Serializer, used for serializing response into
+    other objects.
 
-    __slots__ = ("_response",)
+    Attributes:
+        response (httpx.Response): The response API returned.
+
+    Warning:
+        This object is not for public use unless `await Client.get_raw_response()`
+        is being used.
+    """
+
+    __slots__ = ("response",)
 
     def __init__(self, response: httpx.Response) -> None:
-        self._response = response
+        self.response = response
 
-    async def manifest_content(self) -> Manifest:
+    def manifest_content(self) -> Manifest:
         """
-        Serializes into Manifest.
+        Serializes into [Manifest](./manifest.md).
+
+        Returns:
+            A [Manifest](./manifest.md) object containing mission's info.
         """
-        data = (self._response.json())["rover"]
+        data = (self.response.json())["rover"]
         if data:
             return Manifest(data)
         else:
             raise BadContentError(content=data)
 
-    async def photo_content(self) -> list:
+    def photo_content(self) -> Optional[list]:
         """
-        Serializes into Photo.
+        Serializes into a list of [Photo](./photo.md).
+
+        Returns:
+            A list of [Photo](./photo.md) objects with url and info.
         """
-        data = self._response.json()
-        for key in data:
-            if key in ("photos", "latest_photos"):
-                data = data[key]
-        if data != []:
-            return [marsworks.Photo(img) for img in data]
-        else:
-            return data
+        data = self.response.json()
+        options = ("photos", "latest_photos")
+        if any(option in data for option in options):
+            return [marsworks.Photo(img) for img in data[list(data)[0]]]
+        raise BadContentError(content=data)
 
     def __repr__(self):
-        fil = filter(
-            lambda attr: not attr[0].startswith("_")
-            and not callable(getattr(self, attr[0], None)),
-            inspect.getmembers(self),
-        )
-        rpr = "".join(f"{i[0]} = {i[1]}, " for i in fil)[:-2]
-        return f"{__class__.__name__}({rpr})"
+        attrs = [
+            attr
+            for attr in inspect.getmembers(self)
+            if not inspect.ismethod(attr[1])
+            if not attr[0].startswith("_")
+        ]
+        fmt = ", ".join(f"{attr}={value}" for attr, value in attrs)[:-2]
+        return f"{__class__.__name__}({fmt})"